requirements.txt
atproto
python-dotenv

numpy
pandas
matplotlib
pytz